Move ILocalAppSettings
