--- src/commands/appSettings/ILocalAppSettings.ts
@@ -1,10 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export interface ILocalAppSettings {
-    IsEncrypted?: boolean;
-    Values?: { [key: string]: string };
-    ConnectionStrings?: { [key: string]: string };
-}
